fix all .forward() methods to use only tensors (necessary for pipe parallel)
